alpha-prophet/requirements.txt
opengradient>=0.7.0
tensorflow>=2.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
onnx>=1.15.0
tf2onnx>=1.16.0
plotly>=5.18.0
